backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
python-dotenv
pymongo
requests
beautifulsoup4
